requirements.txt
numpy==1.20.1
av==8.0.3
requests==2.25.1
Pillow==8.1.0
black
python-dotenv
virtualvideo==0.1.6
pyscreenshot==3.0